Fix for OPENAM-362
git-svn-id: https://svn.forgerock.org/openam/trunk@474 0f4defcf-c51a-4c67-9f44-6fb5eba73c5d

+++ opensso/products/amserver/source/com/sun/identity/authentication/distUI/LoginViewBean.java
@@ -473,7 +473,8 @@ extends com.sun.identity.authentication.UI.AuthViewBeanBase {
                     clearCookieAndDestroySession();
                 }
                 try {
-                    if (ssoToken != null) {
+                    if (ssoToken != null &&
+                            SystemProperties.getAsBoolean(Constants.DESTROY_SESSION_AFTER_UPGRADE)) {
                         loginDebug.message(
                             "Destroy existing/old valid session");
                         manager.destroyToken(ssoToken);
